requirements.txt
arxiv==2.1.3
openai==1.35.7
python-dotenv==1.0.1
PyGithub==2.3.0